:recycle: 최상위로 Provider 이동하여 ErrorBoundary 구조 세팅
